# docs/03-api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Card Fraud Ops Analyst Agent",
    "description": "LangGraph-based autonomous fraud analyst assistant. Uses planner-driven tool orchestration for fraud investigation.",
    "version": "0.1.0",
    "x-generated-at": "2026-02-25T03:26:21.594653+00:00"
  },
  "paths": {
...
    "/api/v1/ops-agent/investigations": {
      "get": {
        "tags": [
          "investigations"
        ],
        "summary": "List Investigations",
        "description": "List investigations with optional filters.",
        "operationId": "list_investigations_api_v1_ops_agent_investigations_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "transaction_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Transaction Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvestigationListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ops-agent/investigations/run": {
      "post": {
        "tags": [
          "investigations"
        ],
        "summary": "Run Investigation",
        "description": "Start a new fraud investigation.",
        "operationId": "run_investigation_api_v1_ops_agent_investigations_run_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvestigationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/ops-agent/investigations/{investigation_id}": {
      "get": {
        "tags": [
          "investigations"
        ],
        "summary": "Get Investigation",
        "description": "Get full investigation details.",
        "operationId": "get_investigation_api_v1_ops_agent_investigations__investigation_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "investigation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Investigation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvestigationDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ops-agent/investigations/{investigation_id}/trace": {
      "get": {
        "tags": [
          "investigations"
        ],
        "summary": "Get Investigation Trace",
        "description": "Get HTML trace viewer for investigation.",
        "operationId": "get_investigation_trace_api_v1_ops_agent_investigations__investigation_id__trace_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "investigation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Investigation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ops-agent/investigations/{investigation_id}/resume": {
      "post": {
        "tags": [
          "investigations"
        ],
        "summary": "Resume Investigation",
        "description": "Resume a failed or interrupted investigation.",
        "operationId": "resume_investigation_api_v1_ops_agent_investigations__investigation_id__resume_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "investigation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Investigation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvestigationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ops-agent/investigations/{investigation_id}/rule-draft": {
      "get": {
        "tags": [
          "investigations"
        ],
        "summary": "Get Rule Draft",
        "description": "Get rule draft for an investigation.",
        "operationId": "get_rule_draft_api_v1_ops_agent_investigations__investigation_id__rule_draft_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "investigation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Investigation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "InvestigationDetailResponse": {
        "properties": {
          "investigation_id": {
            "type": "string",
            "title": "Investigation Id"
          },
          "transaction_id": {
            "type": "string",
            "title": "Transaction Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "severity": {
            "type": "string",
            "title": "Severity"
          },
          "confidence_score": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence Score"
          },
          "step_count": {
            "type": "integer",
            "title": "Step Count"
          },
          "max_steps": {
            "type": "integer",
            "title": "Max Steps",
            "default": 20
          },
          "planner_decisions": {
            "items": {
              "$ref": "#/components/schemas/PlannerDecisionSchema"
            },
            "type": "array",
            "title": "Planner Decisions"
          },
          "tool_executions": {
            "items": {
              "$ref": "#/components/schemas/ToolExecutionSchema"
            },
            "type": "array",
            "title": "Tool Executions"
          },
          "recommendations": {
            "items": {
              "$ref": "#/components/schemas/RecommendationSchema"
            },
            "type": "array",
            "title": "Recommendations"
          },
          "started_at": {
            "type": "string",
            "title": "Started At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "total_duration_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Duration Ms"
          },
          "context": {
            "additionalProperties": true,
            "type": "object",
            "title": "Context"
          },
          "evidence": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Evidence"
          },
          "pattern_results": {
            "additionalProperties": true,
            "type": "object",
            "title": "Pattern Results"
          },
          "similarity_results": {
            "additionalProperties": true,
            "type": "object",
            "title": "Similarity Results"
          },
          "reasoning": {
            "additionalProperties": true,
            "type": "object",
            "title": "Reasoning"
          },
          "hypotheses": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Hypotheses"
          },
          "rule_draft": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rule Draft"
          }
        },
        "type": "object",
        "required": [
          "investigation_id",
          "transaction_id",
          "status",
          "severity",
          "confidence_score",
          "step_count",
          "started_at"
        ],
        "title": "InvestigationDetailResponse"
      },
      "InvestigationListResponse": {
        "properties": {
          "investigations": {
            "items": {
              "$ref": "#/components/schemas/InvestigationSummary"
            },
            "type": "array",
            "title": "Investigations"
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "investigations"
        ],
        "title": "InvestigationListResponse"
      },
      "InvestigationResponse": {
        "properties": {
          "investigation_id": {
            "type": "string",
            "title": "Investigation Id"
          },
          "transaction_id": {
            "type": "string",
            "title": "Transaction Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "severity": {
            "type": "string",
            "title": "Severity"
          },
          "confidence_score": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence Score"
          },
          "step_count": {
            "type": "integer",
            "title": "Step Count"
          },
          "max_steps": {
            "type": "integer",
            "title": "Max Steps",
            "default": 20
          },
          "planner_decisions": {
            "items": {
              "$ref": "#/components/schemas/PlannerDecisionSchema"
            },
            "type": "array",
            "title": "Planner Decisions"
          },
          "tool_executions": {
            "items": {
              "$ref": "#/components/schemas/ToolExecutionSchema"
            },
            "type": "array",
            "title": "Tool Executions"
          },
          "recommendations": {
            "items": {
              "$ref": "#/components/schemas/RecommendationSchema"
            },
            "type": "array",
            "title": "Recommendations"
          },
          "started_at": {
            "type": "string",
            "title": "Started At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "total_duration_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Duration Ms"
          }
        },
        "type": "object",
        "required": [
          "investigation_id",
          "transaction_id",
          "status",
          "severity",
          "confidence_score",
          "step_count",
          "started_at"
        ],
        "title": "InvestigationResponse"
      },
      "InvestigationSummary": {
        "properties": {
          "investigation_id": {
            "type": "string",
            "title": "Investigation Id"
          },
          "transaction_id": {
            "type": "string",
            "title": "Transaction Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "severity": {
            "type": "string",
            "title": "Severity"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          }
        },
        "type": "object",
        "required": [
          "investigation_id",
          "transaction_id",
          "status",
          "severity"
        ],
        "title": "InvestigationSummary"
      },
...
      "PlannerDecisionSchema": {
        "properties": {
          "step": {
            "type": "integer",
            "title": "Step"
          },
          "selected_tool": {
            "type": "string",
            "title": "Selected Tool"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "confidence": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence"
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp"
          }
        },
        "type": "object",
        "required": [
          "step",
          "selected_tool",
          "reason",
          "confidence",
          "timestamp"
        ],
        "title": "PlannerDecisionSchema"
      },
...
      "RecommendationSchema": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type"
          },
          "priority": {
            "type": "integer",
            "title": "Priority",
            "default": 0
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "impact": {
            "type": "string",
            "title": "Impact"
          },
          "signature_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Signature Hash"
          }
        },
        "type": "object",
        "required": [
          "type",
          "title",
          "impact"
        ],
        "title": "RecommendationSchema"
      },
...
      "RunRequest": {
        "properties": {
          "transaction_id": {
            "type": "string",
            "minLength": 1,
            "title": "Transaction Id",
            "description": "Transaction UUID"
          },
          "mode": {
            "type": "string",
            "title": "Mode",
            "default": "FULL"
          }
        },
        "type": "object",
        "required": [
          "transaction_id"
        ],
        "title": "RunRequest"
      },
...
      "ToolExecutionSchema": {
        "properties": {
          "tool_name": {
            "type": "string",
            "title": "Tool Name"
          },
          "input_summary": {
            "additionalProperties": true,
            "type": "object",
            "title": "Input Summary"
          },
          "output_summary": {
            "additionalProperties": true,
            "type": "object",
            "title": "Output Summary"
          },
          "execution_time_ms": {
            "type": "integer",
            "title": "Execution Time Ms"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp"
          }
        },
        "type": "object",
        "required": [
          "tool_name",
          "execution_time_ms",
          "status",
          "timestamp"
        ],
        "title": "ToolExecutionSchema"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}